+++ backend/src/main/java/Testing/UserResource.java
@@ -30,31 +30,4 @@ public class UserResource {
     public void createUser(User user) {
         em.persist(user);
     }
-
-    @POST
-    @Path("/login")
-    @Transactional
-    public Response login(User request) {
-        try {
-            User user = em.createQuery("SELECT u FROM User u WHERE u.email = :email", User.class)
-                    .setParameter("email", request.getEmail())
-                    .getSingleResult();
-
-            if (!user.getPassword().equals(request.getPassword())) {
-                return Response.status(Response.Status.UNAUTHORIZED)
-                        .entity("Invalid password")
-                        .build();
-            }
-
-
-            String token = JwtUtil.generateToken(user.getEmail(), user.getRole());
-
-            return Response.ok().entity("{\"token\": \"" + token + "\"}").build();
-        } catch (NoResultException e) {
-            return Response.status(Response.Status.UNAUTHORIZED)
-                    .entity("User not found")
-                    .build();
-        }
-    }
-
 }
