+++ backend/src/main/resources/application.properties
@@ -2,9 +2,9 @@
 quarkus.smallrye-jwt.enabled=true
 
 # Signing JWT
-#smallrye.jwt.sign.key-location=classpath:privatePem.key
 smallrye.jwt.sign.key-location=classpath:privateKey-pkcs8.pem
 mp.jwt.verify.publickey.location=classpath:publicKey.pem
+
 
 
 smallrye.jwt.sign.key-format=PEM
@@ -12,7 +12,6 @@ smallrye.jwt.sign.algorithm=RS256
 
 # Verification JWT
 mp.jwt.verify.issuer=https://example.com/issuer
-#mp.jwt.verify.publickey.location=classpath:META-INF/resources/publicKey.pem
 
 # Database
 quarkus.datasource.db-kind=postgresql
@@ -26,3 +25,9 @@ quarkus.hibernate-orm.database.generation=update
 quarkus.http.cors=true
 quarkus.http.cors.origins=http://localhost:5173
 quarkus.http.cors.headers=Authorization,Content-Type
+
+quarkus.log.level=DEBUG
+quarkus.log.category."io.smallrye.jwt".level=DEBUG
+quarkus.log.category."io.smallrye.jwt.build.impl".level=DEBUG
+quarkus.log.category."io.quarkus".level=DEBUG
+
