+++ backend/src/main/resources/publicKey.pem
@@ -1,9 +1,9 @@
 -----BEGIN PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAvW4ZndClkGknTF3JGZI8
-Qe2VwJZIcM7o/Okh1EnNw/9rAOg2WyBRl//ILJtFutymhtYTcbEFke6Gam+n2Ck+
-dqHM3RW8NKuimPMtYTZ4D7J7zPBZqUW6B18ju3HaugNli/IAoMlwqWl2oxvMA5D3
-RKx4hXGsNt9gxDQVYNbDhmOZJtrnbY8by1Hm6IUf+Kl265LE7MMSanNaNzVHgSGD
-gO1dn+xg3Rm2MZk0eIUKZaeOGzXw+CmAlZqDnVv0HRcgcXVodY0jdyvjSn0Jg3Ut
-pVakOqGLbm+s2gPjKJfQBLHVrFwNSg91jAF4XwP3eF/F3NUCjfRgAsYthTvZN2ce
-WQIDAQAB
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAm2Pu1rfRddromGzxydMB
+nwWnQaZ5466ImjCA+2ODR7KveVRq85w9R3Il6Ts39DGlZC5HCl3L393Y8NSH3BEn
+CQg+jaKbVUkD0jIdCYubkfcO7KDJmaEAQysWCtsTUx1NRSYjh1qxTRO8GXTeZ2W0
+ok2n5fkF/Xw2YUNLFsABriYmkv031dhzDPL7F3CtEGRZicQe9pKgXlAUctuHMIxY
+0b8FNjk1f4IDlb+I4e53iWGK/iPOpfUWegpqpi0hrITBeNQafEm6c0bb/S0qN+bW
+UWtxxDAhLB+QOn6l78gVr4uelwUHK/z3HRhmfP0Gw8CeOC+erxqrCkoRR6/4iLz/
+4wIDAQAB
 -----END PUBLIC KEY-----
